refactor(sdf): separate out `resources.py`
